Created MeleeEnemy Class
Created the melee enemy class and switched the current spawner to spawn it instead. Currently it uses the state machine system I was using before creating the basic enemy. The follow player function in the enemy class is not working as expected, but other than that it switches states properly. Need to add more states.